Moved general components styles to general folder
