+++ src/assets/styles/components/general/carousel.scss
