+++ src/assets/styles/components/general/footer.scss
@@ -1,4 +1,4 @@
-@import '../global-style/styles.scss';
+@import '../../global-style/styles.scss';
 
 footer {
   height: $footer-h;
