+++ src/assets/styles/components/general/header.scss
@@ -1,4 +1,4 @@
-@import '../global-style/styles.scss';
+@import '../../global-style/styles.scss';
 
 header{
     height: $header-h;
