+++ src/assets/styles/components/general/search.scss
@@ -1,4 +1,4 @@
-@import '../global-style/styles.scss';
+@import '../../global-style/styles.scss';
 
 .icon {
   margin: 0 1rem;
