+++ src/components/general/Card.tsx
@@ -1,8 +1,12 @@
-import "../../assets/styles/components/card.scss";
+import "../../assets/styles/components/general/card.scss";
 import React, { FC } from "react";
 import CardProps from "../../utils/interfaces/card";
+import { isTitleTooLong } from "../../utils/functions/functions";
 
 const CardComponent: FC<CardProps> = (props: CardProps) => {
+
+  const changeTitleFontSize = (title: string)=> isTitleTooLong(title, 10) ? {fontSize: '1.6rem'} : {}
+
   const MediumCard = () => <h3>{props.type === "medium" && props.subTitle}</h3>;
   const LargeCard = () => (
     <>
@@ -27,7 +31,7 @@ const CardComponent: FC<CardProps> = (props: CardProps) => {
       }`}
     >
       <img src={props.picture} alt="props.title" />
-      <h2>{props.title}</h2>
+      <h2 style={changeTitleFontSize(props.title)}>{props.title}</h2>
       {props.type === "medium" && <MediumCard />}
       {props.type === "large" && <LargeCard />}
     </div>
